requirements.txt
aiogram==3.4.1
openai==1.30.1
gTTS==2.5.1
python-dotenv==1.0.1
